19/10 - 16:50
pelo menos não ta dando erro no terminal

+++ Back-End-Maven/almoxarifado/src/main/java/br/senai/sc/almoxarifado/model/entities/SacolaProduto.java
@@ -14,6 +14,11 @@ import javax.persistence.*;
 @EqualsAndHashCode
 public class SacolaProduto {
     // Foreign keys
+
+//    @EmbeddedId
+//    public SacolaProdutoKey getId() {
+//        return id;
+//    }
     @EmbeddedId
     SacolaProdutoKey id;
 
@@ -30,5 +35,5 @@ public class SacolaProduto {
     // Attribute
 
     @Column(nullable = false)
-    private Integer qtd_produto;
+    private Long qtd_produto;
 }
